backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
mysqlclient==2.2.0
Pillow==10.1.0
python-decouple==3.8
reportlab==4.0.7
openpyxl==3.1.2
django-crispy-forms==2.1
crispy-bootstrap5==0.7
